Prepare safe fleet load SDK surface
Extract fleet config reads into a core-only helper so broadcast can later consume SDK exports without pulling the shared fleet-load module cycle.

Constraint: #2156 asks for a safer fleet-load-core variant before switching broadcast imports to @maw-js/sdk.

Confidence: medium

Scope-risk: narrow

Tested: not run; interrupted for urgent #2163 rebase.

Not-tested: SDK exports and broadcast migration still pending.
