+++ src/commands/shared/fleet-load.ts
@@ -1,139 +1,40 @@
-import { join } from "path";
-import { existsSync, readdirSync, readFileSync } from "fs";
 import * as sdk from "../../sdk";
-import { fleetDirForWrite as coreFleetDirForWrite, fleetDirsForRead as coreFleetDirsForRead, uniqueDirs } from "../../core/fleet/paths";
+import {
+  fleetDirForWrite,
+  fleetDirsForRead as coreFleetDirsForRead,
+  loadFleet as loadFleetCore,
+  countDisabledFleetFiles as countDisabledFleetFilesCore,
+  loadDisabledFleetEntries as loadDisabledFleetEntriesCore,
+  loadFleetEntries as loadFleetEntriesCore,
+  type DisabledFleetEntry,
+  type FleetEntry,
+  type FleetSession,
+  type FleetWindow,
+} from "../../core/fleet/fleet-load-core";
 import { resolveFleetWindowSessionTarget } from "../../core/matcher/resolve-target";
 
-export interface FleetWindow {
-  name: string;
-  repo: string;
-}
-
-export interface FleetSession {
-  name: string;
-  windows: FleetWindow[];
-  skip_command?: boolean;
-  /** Peer oracle names for soul-sync (flat, no hierarchy). */
-  sync_peers?: string[];
-  /** Optional parent oracle/fleet name for bud lineage. */
-  budded_from?: string;
-  /** Project repos (org/repo) this oracle absorbs ψ/ from via `maw soul-sync --project`. */
-  project_repos?: string[];
-}
-
-export interface FleetEntry {
-  file: string;
-  /** Absolute path of the config file that supplied this entry. */
-  path?: string;
-  num: number;
-  groupName: string;
-  session: FleetSession;
-}
-
-export interface DisabledFleetEntry {
-  file: string;
-  /** Absolute path of the disabled config file that supplied this entry. */
-  path: string;
-  num: number;
-  groupName: string;
-  session?: FleetSession;
-  error?: unknown;
-}
+export type { DisabledFleetEntry, FleetEntry, FleetSession, FleetWindow };
 
 export function fleetDirsForRead(): string[] {
-  const legacyFleetDir = (sdk as any).FLEET_DIR as string | undefined;
-  return legacyFleetDir ? coreFleetDirsForRead({ legacyFleetDir }) : uniqueDirs([coreFleetDirForWrite()]);
+  return coreFleetDirsForRead((sdk as any).FLEET_DIR as string | undefined);
 }
 
-export function fleetDirForWrite(): string {
-  return coreFleetDirForWrite();
-}
-
-function readFleetFiles(dirs: string[] = fleetDirsForRead()): Array<{ file: string; path: string; session: FleetSession }> {
-  const byName = new Map<string, { file: string; path: string; session: FleetSession }>();
-  for (const dir of uniqueDirs(dirs)) {
-    if (!existsSync(dir)) continue;
-    let files: string[];
-    try {
-      files = readdirSync(dir)
-        .filter(f => f.endsWith(".json") && !f.endsWith(".disabled"))
-        .sort();
-    } catch {
-      continue;
-    }
-    for (const file of files) {
-      if (byName.has(file)) continue;
-      const path = join(dir, file);
-      try {
-        byName.set(file, { file, path, session: JSON.parse(readFileSync(path, "utf-8")) as FleetSession });
-      } catch {
-        // Keep looking in fallback dirs. A malformed state-first fleet file
-        // should not shadow a valid legacy config with the same filename.
-      }
-    }
-  }
-  return [...byName.values()].sort((a, b) => a.file.localeCompare(b.file));
-}
-
-function readDisabledFleetFiles(dirs: string[] = fleetDirsForRead()): Array<{ file: string; path: string }> {
-  const byName = new Map<string, { file: string; path: string }>();
-  for (const dir of uniqueDirs(dirs)) {
-    if (!existsSync(dir)) continue;
-    let files: string[];
-    try {
-      files = readdirSync(dir)
-        .filter(f => f.endsWith(".disabled"))
-        .sort();
-    } catch {
-      continue;
-    }
-    for (const file of files) {
-      if (!byName.has(file)) byName.set(file, { file, path: join(dir, file) });
-    }
-  }
-  return [...byName.values()].sort((a, b) => a.file.localeCompare(b.file));
-}
-
-function parseFleetFileInfo(file: string): { num: number; groupName: string } {
-  const activeName = file.replace(/\.disabled$/i, "");
-  const match = activeName.match(/^(\d+)-(.+)\.json$/);
-  return {
-    num: match ? parseInt(match[1], 10) : 0,
-    groupName: match ? match[2] : activeName.replace(/\.json$/i, ""),
-  };
-}
+export { fleetDirForWrite };
 
 export function loadFleet(dirs: string[] = fleetDirsForRead()): FleetSession[] {
-  return readFleetFiles(dirs).map(({ session }) => session);
+  return loadFleetCore(dirs);
 }
 
-
 export function countDisabledFleetFiles(dirs: string[] = fleetDirsForRead()): number {
-  return readDisabledFleetFiles(dirs).length;
+  return countDisabledFleetFilesCore(dirs);
 }
 
 export function loadDisabledFleetEntries(dirs: string[] = fleetDirsForRead()): DisabledFleetEntry[] {
-  return readDisabledFleetFiles(dirs).map(({ file, path }) => {
-    const { num, groupName } = parseFleetFileInfo(file);
-    try {
-      return { file, path, num, groupName, session: JSON.parse(readFileSync(path, "utf-8")) as FleetSession };
-    } catch (error) {
-      return { file, path, num, groupName, error };
-    }
-  });
+  return loadDisabledFleetEntriesCore(dirs);
 }
 
 export function loadFleetEntries(dirs: string[] = fleetDirsForRead()): FleetEntry[] {
-  return readFleetFiles(dirs).map(({ file, path, session }) => {
-    const { num, groupName } = parseFleetFileInfo(file);
-    return {
-      file,
-      path,
-      num,
-      groupName,
-      session,
-    };
-  });
+  return loadFleetEntriesCore(dirs);
 }
 
 export async function getSessionNames(): Promise<string[]> {
